Adding a verification for Google Search Console
Seems weird that they would use an HTML file and not another format but this is one of the site ownership verification methods provided by Google directly, so I prostrate myself before the mighty gods of SEO and submit this sacrificial lamb. May it please them well.

If it doesn't work for any reason I will take it down.